package structure change, Layer refactor
