+++ props-aws/src/main/java/sh/props/aws/AwsSecretsManagerOnDemand.java
@@ -34,8 +34,8 @@ import java.util.List;
 import java.util.Random;
 import java.util.concurrent.CompletableFuture;
 import java.util.stream.Stream;
-import sh.props.source.OnDemandSource;
-import sh.props.source.Source;
+import sh.props.OnDemandSource;
+import sh.props.Source;
 import software.amazon.awssdk.core.client.config.ClientOverrideConfiguration;
 import software.amazon.awssdk.regions.Region;
 import software.amazon.awssdk.services.secretsmanager.SecretsManagerAsyncClient;
