+++ props-core/src/main/java/sh/props/CustomPropBuilder.java
@@ -25,7 +25,7 @@
 
 package sh.props;
 
-import static sh.props.Registry.assertNotNull;
+import static sh.props.util.Validate.assertNotNull;
 
 import sh.props.annotations.Nullable;
 import sh.props.converter.Converter;
