+++ props-core/src/main/java/sh/props/OnDemandSource.java
@@ -23,7 +23,7 @@
  *
  */
 
-package sh.props.source;
+package sh.props;
 
 import static java.lang.String.format;
