+++ props-core/src/main/java/sh/props/Scheduler.java
@@ -23,14 +23,13 @@
  *
  */
 
-package sh.props.source.refresh;
+package sh.props;
 
 import static java.util.concurrent.TimeUnit.NANOSECONDS;
 
 import java.time.Duration;
 import java.util.concurrent.ForkJoinPool;
 import java.util.concurrent.ScheduledExecutorService;
-import sh.props.source.Source;
 import sh.props.util.BackgroundExecutorFactory;
 
 public class Scheduler {
