+++ props-core/src/test/java/sh/props/source/impl/RefreshedSourcesTest.java
@@ -37,10 +37,11 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.time.Duration;
 import org.junit.jupiter.api.Test;
+import sh.props.FileWatchSvc;
 import sh.props.RegistryBuilder;
+import sh.props.Scheduler;
 import sh.props.converter.Cast;
-import sh.props.source.refresh.FileWatchSvc;
-import sh.props.source.refresh.Scheduler;
+import sh.props.source.PropertyFile;
 import sh.props.textfixtures.AwaitAssertionTest;
 import sh.props.textfixtures.TestFileUtil;
 import sh.props.typed.BooleanProp;
