+++ props-core/src/test/java/sh/props/source/impl/SourcesTest.java
@@ -30,8 +30,8 @@ import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.notNullValue;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.IsNull.nullValue;
-import static sh.props.source.impl.InMemory.UPDATE_REGISTRY_MANUALLY;
-import static sh.props.source.impl.InMemory.UPDATE_REGISTRY_ON_EVERY_WRITE;
+import static sh.props.source.InMemory.UPDATE_REGISTRY_MANUALLY;
+import static sh.props.source.InMemory.UPDATE_REGISTRY_ON_EVERY_WRITE;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -41,6 +41,11 @@ import org.junit.jupiter.api.Assumptions;
 import org.junit.jupiter.api.Test;
 import sh.props.RegistryBuilder;
 import sh.props.converter.Cast;
+import sh.props.source.ClasspathPropertyFile;
+import sh.props.source.Environment;
+import sh.props.source.InMemory;
+import sh.props.source.PropertyFile;
+import sh.props.source.SystemProperties;
 import sh.props.textfixtures.AwaitAssertionTest;
 import sh.props.textfixtures.TestFileUtil;
 import sh.props.typed.BooleanProp;
